Remove StatusSala property from Sala entity
This commit introduces a migration class `RemovidoStatusSalaDaEntidadeSala` that removes the `StatusSala` column from the `TAB_Sala` table. The `Up` method drops the column, while the `Down` method allows for its restoration if needed. The `RoomFlowContextModelSnapshot.cs` file has also been updated to reflect this change, ensuring the database schema is in sync with the updated entity model.